281: create new /think file structure
