+++ src/think/fsm.js
@@ -1,5 +1,4 @@
 import _ from 'lodash';
-import { getPointsAlongPath } from '../utils/fsmUtils';
 import { getDist } from '../utils/geometry';
 import { drawEnemyPolypointPath } from '../draw/triangulation';
 import { amRed,
@@ -13,6 +12,39 @@ import { findCachedTile, findEnemyFC } from '../look/locations';
 import { BRP } from '../constants';
 import { getShortestPolypointPath } from '../navmesh/path';
 import { getDTGraph } from '../navmesh/triangulation';
+
+
+/**
+ * @param {{point: {x: number, y: number}}[]} path - a list of objects with a point key.
+ * @param {number} granularity - the euclidian distance the points along the path should be
+ *   separated by.
+ * @returns {{x: number, y: number}[]} a list of points along the path. These points result from
+ *   connecting each point in the input path, then placing points along each line, separated by
+ *   granularity
+ */
+function getPointsAlongPath(path, granularity = 40) {
+  const res = [];
+  for (let i = 0; i < path.length - 1; i += 1) {
+    const currPoint = path[i].point;
+    const nextPoint = path[i + 1].point;
+    const edgeLength = getDist(
+      currPoint.x,
+      currPoint.y,
+      nextPoint.x,
+      nextPoint.y,
+    );
+    const increments = Math.floor(edgeLength / granularity);
+    const xIncr = (nextPoint.x - currPoint.x) / increments;
+    const yIncr = (nextPoint.y - currPoint.y) / increments;
+    res.push(currPoint);
+    for (let j = 0; j < increments - 1; j += 1) {
+      const prev = _.last(res);
+      res.push({ x: prev.x + xIncr, y: prev.y + yIncr });
+    }
+  }
+  res.push(_.last(path));
+  return res;
+}
 
 
 export function chaseEnemyFC(me, goal, enemyFC, enemyShortestPath) {
